FEAT: Self-host Streaks Stat App on Vercel
Note that this self-host uses a snapshot of the original repository
instead of a fork, and therefore does not update from this point.
If updates are necessary, we will have to consider forking from the
original repository and using the `vercel` branch as the main deployment
branch. This will require additional tweaks, and is untested.